clean up settings menu
